Split flags into new file
